+++ src/Printed/Common/PdfTools/Cpdf/CpdfPdfSplitter.php
@@ -2,26 +2,44 @@
 
 namespace Printed\Common\PdfTools\Cpdf;
 
+use Printed\Common\Filesystem\TemporaryFile;
 use RuntimeException;
+use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\Process\Process;
 
 class CpdfPdfSplitter
 {
+    /** @var Filesystem */
+    private $filesystem;
+
+    /** @var CpdfPdfInformationExtractor */
+    private $cpdfPdfInformationExtractor;
+
     /** @var CpdfBinaryConfiguration */
     private $binaryConfig;
 
     /**
+     * @param Filesystem $filesystem
+     * @param CpdfPdfInformationExtractor $cpdfPdfInformationExtractor
      * @param string $binaryPath Full path to the cpdf binary.
      *
      * @throws Exception\CpdfException
      */
-    public function __construct($binaryPath)
-    {
+    public function __construct(
+        Filesystem $filesystem,
+        CpdfPdfInformationExtractor $cpdfPdfInformationExtractor,
+        $binaryPath
+    ) {
+        $this->filesystem = $filesystem;
+        $this->cpdfPdfInformationExtractor = $cpdfPdfInformationExtractor;
         $this->binaryConfig = CpdfBinaryConfiguration::create($binaryPath);
     }
 
     /**
+     * The split files will be put in the same directory as the source file. This might cause problems
+     * (i.e. name conflicts) if you misuse this method.
+     *
      * @param File $pdfFile
      * @param array $options An array of options to use in the spit command.
      *
@@ -30,7 +48,18 @@ class CpdfPdfSplitter
     public function split(File $pdfFile, array $options = [])
     {
         $options = array_merge([
+            /*
+             * @deprecated This option now is always on due to cpdf "-remove-bookmarks" op's creating broken files without
+             *  it. The cases that are fixed by having this option always on are unit tested, so feel free to disable this
+             *  option to see in what way the pdfs are broken.
+             */
             'preventPreserveObjectStreams' => true,
+            /*
+             * @var int|null Max number of extracted pages. Extra pages are ignored. You most likely always want to somehow
+             *  ensure that you don't just split any pdf the end user provides you with. This options is one of the ways.
+             *  Null means "no limit".
+             */
+            'maxPageCount' => null,
         ], $options);
 
         if ($pdfFile->guessExtension() !== 'pdf') {
@@ -45,10 +74,6 @@ class CpdfPdfSplitter
 
         $extraArguments = [];
 
-        if ($options['preventPreserveObjectStreams']) {
-            $extraArguments[] = '-no-preserve-objstm';
-        }
-
         $outputPathname = sprintf(
             '%s/%s.@N.%s',
             $pathInfo['dirname'],
@@ -57,11 +82,39 @@ class CpdfPdfSplitter
         );
 
         /*
+         * Max page count. The number must not be larger than the number of pages in the pdf.
+         */
+        $pageCountToExtract = $this->cpdfPdfInformationExtractor->getPagesCount($pdfFile);
+        if (
+            $options['maxPageCount']
+            && $pageCountToExtract > $options['maxPageCount']
+        ) {
+            $pageCountToExtract = $options['maxPageCount'];
+        }
+
+        /*
+         * PDF splitting is done in 2 steps:
+         *
+         * 1. Remove bookmarks. Save output to a temporary file.
+         * 2. Split the pdf stored in the temporary file.
+         *
+         * Reason: to cover all use cases. This is unit tested to feel free to experiment with different implementations
+         * if needed.
+         */
+
+        /*
          * -remove-bookmarks fixes an issue caused by corrupt bookmarks.
          * @see https://github.com/johnwhitington/cpdf-source/issues/123
          */
+        $intermediaryTemporaryFile = $this->removePdfBookmarks($pdfFile);
+
         $command = implode(' ', [
-            sprintf('exec ./%s -remove-bookmarks -i %s', $this->binaryConfig->getFilename(), $inputPathname),
+            sprintf(
+                'exec ./%1$s -i %2$s 1-%3$d',
+                $this->binaryConfig->getFilename(),
+                $intermediaryTemporaryFile->getPathname(),
+                $pageCountToExtract
+            ),
             sprintf(
                 'AND %s -split -o %s',
                 implode(' ', $extraArguments),
@@ -88,5 +141,32 @@ class CpdfPdfSplitter
         natsort($outputFiles);
 
         return array_values($outputFiles);
+    }
+
+    /**
+     * @param File $pdfFile
+     * @return TemporaryFile
+     */
+    private function removePdfBookmarks(File $pdfFile)
+    {
+        /*
+         * Future improvement: Better way would be to use a factory for temporary files but there wasn't available at the
+         * time of writing this. Apologies for #excuses.
+         *
+         * Note that local /tmp folder is preferred here (over any mounted equivalent) for speed.
+         */
+        $intermediaryTemporaryFile = new TemporaryFile($this->filesystem, sys_get_temp_dir());
+
+        $cliCommand = sprintf(
+            'exec ./%s -remove-bookmarks -i %s AND -no-preserve-objstm -o %s',
+            $this->binaryConfig->getFilename(),
+            $pdfFile->getPathname(),
+            $intermediaryTemporaryFile->getPathname()
+        );
+
+        $removePdfBookmarksProcess = new Process($cliCommand, $this->binaryConfig->getPath());
+        $removePdfBookmarksProcess->mustRun();
+
+        return $intermediaryTemporaryFile;
     }
 }