+++ src/Printed/Common/PdfTools/Tests/TestUtils.php
@@ -2,6 +2,8 @@
 
 namespace Printed\Common\PdfTools\Tests;
 
+use Symfony\Component\HttpFoundation\File\File;
+
 class TestUtils
 {
     /**
@@ -9,7 +11,12 @@ class TestUtils
      */
     public static function getProjectDir()
     {
-        return realpath(__DIR__ . '/../../../../..');
+        static $projectDir = null;
+        if ($projectDir) {
+            return $projectDir;
+        }
+
+        return $projectDir = realpath(__DIR__ . '/../../../../..');
     }
 
     /**
@@ -20,5 +27,21 @@ class TestUtils
         $binaryName = getenv('ALPINE') ? 'cpdf-alpine' : 'cpdf';
 
         return sprintf('%s/vendor/bin/%s', self::getProjectDir(), $binaryName);
+    }
+
+    /**
+     * Get a test file from "printed/common-test-files"
+     *
+     * @param string $testFileName
+     * @return File
+     */
+    public static function getPrintedCommonTestFile($testFileName)
+    {
+        $projectDir = self::getProjectDir();
+
+        /*
+         * I'm not amazed with the hardcoded composer vendor subdir here. But you know, it worked at the time.
+         */
+        return new File("{$projectDir}/vendor/printed/common-test-files/pdf/{$testFileName}");
     }
 }