+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0]
+### Changed
+- [CpdfPdfSplitter] `maxPageCount` option: limits the number of split pages.
+- (minor breaking change) CpdfPdfSplitter requires more constructor args. 
+
+## [0.1.0]
 ### Changed
 - A full path to the CPDF binary is now required. This is a breaking change.